pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyr"
version = "0.1.0"
description = "AI-native narrative game generator"
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.5.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pyinstaller>=6.0.0",
    "jsonschema>=4.0.0",
]

[project.scripts]
pyr = "pyr.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["pyr"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
